Isolated upload-image, specify folders for upload
